don't treat a short read in a hash job as a fatal error
The eof error code will still be communicated to the handler so this
doesn't change behavior with the current code. This will become
important with v2 support where we may be doing multiple hash operations
in the same hash job. If the first hash op gets an eof we still want to
perform the second hash to keep the results consistent.